UPD: make optimizer global.
Fix optimizer class such that it uses global gradients of the image, and it runs optimization for patches and updates them inside `optimize` function. Fix the optimizer test to be compatible with new Optimizer class.

Experiments with optimizer test showed that the optimization process doesn't converge for large shifts and angles of warping. And the gradients need to be blurred.